- current version also compiles (with few warnings) on   - alpha (Debian)                  [few warnings]   - amd64 (suse)                    [few warnings]   - SunOS (5.9, SUNW,Ultra-60)      [without warning]   - Mac OSX (Darwin 6.8, PowerPC)   - FreeBSD 4.10 (i386)   - NetBSD 1.6.1 (i386) - added support for NetBSD to automake script os.m4 - fixed some bugs in Makefiles:   - when ncurses is not available (was assuming if     ncurses is not there it must be Windoze. Well, that's not the case ;-)   - there must be no blank between "-D" and the macro to define via     "-DSOMETHING", otherwise some MacOSX bails out...
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@601 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ admin/mklistdoc.c
@@ -124,7 +124,7 @@ int scanLine(MACRO_DEF *macroDef,
     /* found something, search for opening bracket */
     macroBegin=p;
     p+=strlen(macroDef->macroName);
-    while(*p && isspace(*p)) p++;
+    while(*p && isspace((int)*p)) p++;
     if (*p && *p!='(')
       return -1;
 
@@ -140,7 +140,7 @@ int scanLine(MACRO_DEF *macroDef,
 
         /* found it, now read the typename */
         p++;
-        while(*p && isspace(*p)) p++;
+        while(*p && isspace((int)*p)) p++;
         typeName=p;
 
         /* find comma */
@@ -150,7 +150,7 @@ int scanLine(MACRO_DEF *macroDef,
             if (!typeNameEnd)
               typeNameEnd=p;
             p++;
-            while(*p && isspace(*p)) p++;
+            while(*p && isspace((int)*p)) p++;
             typePrefix=p;
 
             /* find closing bracket */
@@ -167,16 +167,16 @@ int scanLine(MACRO_DEF *macroDef,
 
                 /* check whether this is a definition */
                 s=buffer;
-                while(*s && isspace(*s)) s++;
+                while(*s && isspace((int)*s)) s++;
                 if (*s=='#') {
                   s++;
                   /* preprocessor command, check for define */
-                  while(*s && isspace(*s)) s++;
+                  while(*s && isspace((int)*s)) s++;
                   if (strncasecmp(s, "define ", 7)==0) {
                     s+=7;
                     /* it is a define, now check if the next nonblank is
                      * the beginning of this macro */
-                    while(*s && isspace(*s)) s++;
+                    while(*s && isspace((int)*s)) s++;
                     if (s==macroBegin) {
                       if (verbosity>1)
                         fprintf(stderr, "Found definition for macro \"%s\".\n",
@@ -217,7 +217,7 @@ int scanLine(MACRO_DEF *macroDef,
                 }
                 return 0;
               }
-              else if (isspace(*p)) {
+              else if (isspace((int)*p)) {
                 if (!typePrefixEnd)
                   typePrefixEnd=p;
               }
@@ -231,7 +231,7 @@ int scanLine(MACRO_DEF *macroDef,
             /* closing bracket missing, probably need more data... */
             return 1;
           }
-          else if (isspace(*p)) {
+          else if (isspace((int)*p)) {
             if (!typeNameEnd)
               typeNameEnd=p;
           }
